refactor/feed_article_repository_test.goの内容をモジュール化しリファクタリング
